backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
asyncpg>=0.29.0  # Async PostgreSQL adapter

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI & ML
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
anthropic>=0.7.0

# Vector Database
weaviate-client>=3.25.0

# Caching & Session
redis>=5.0.0
python-redis>=5.0.0

# HTTP & Requests
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0
celery>=5.3.0  # Background tasks
python-slugify>=8.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Monitoring & Logging
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0 